+++ cycles-ledger/src/main.rs
@@ -100,7 +100,7 @@ fn icrc1_balance_of(account: Account) -> Nat {
 
 fn validate_memo(memo: &Option<Memo>) {
     if let Some(memo) = memo {
-        if memo.0.len() as u64 > config::MAX_MEMO_LENGTH as u64 {
+        if memo.0.len() as u64 > config::MAX_MEMO_LENGTH {
             ic_cdk::trap(&format!(
                 "memo length exceeds the maximum of {} bytes",
                 config::MAX_MEMO_LENGTH,
